+++ Sources/Core/Internal/ExtensionsContainer.swift
@@ -2,7 +2,7 @@
 //  ExtensionsContainer.swift
 //  DITranquillity
 //
-//  Created by Ивлев Александр Евгеньевич on 17/08/2018.
+//  Created by Alexander Ivlev on 17/08/2018.
 //  Copyright © 2018 Alexander Ivlev. All rights reserved.
 //