+++ Sources/Core/Internal/Resolver.swift
@@ -6,41 +6,43 @@
 //  Copyright © 2016 Alexander Ivlev. All rights reserved.
 //
 
+import SwiftLazy
+
 class Resolver {
 
   init(container: DIContainer) {
     self.container = container // unowned
   }
   
-  func resolve<T>(type: T.Type = T.self, name: String? = nil, from framework: DIFramework.Type? = nil) -> T {
+  func resolve<T>(type: T.Type = T.self, name: String? = nil, from framework: DIFramework.Type? = nil, arguments: AnyArguments? = nil) -> T {
     let pType = ParsedType(type: type)
     log(.verbose, msg: "Begin resolve \(description(type: pType))", brace: .begin)
     defer { log(.verbose, msg: "End resolve \(description(type: pType))", brace: .end) }
     
-    return gmake(by: make(by: pType, with: name, from: framework, use: nil))
+    return gmake(by: make(by: pType, with: name, from: framework, use: nil, arguments: arguments))
   }
   
-  func injection<T>(obj: T, from framework: DIFramework.Type? = nil) {
+  func injection<T>(obj: T, from framework: DIFramework.Type? = nil, arguments: AnyArguments? = nil) {
     log(.verbose, msg: "Begin injection in obj: \(obj)", brace: .begin)
     defer { log(.verbose, msg: "End injection in obj: \(obj)", brace: .end) }
 
-    _ = make(by: ParsedType(obj: obj), with: nil, from: framework, use: obj)
+    _ = make(by: ParsedType(obj: obj), with: nil, from: framework, use: obj, arguments: arguments)
   }
 
   
-  func resolveCached(component: Component) {
+  func resolveCached(component: Component, arguments: AnyArguments? = nil) {
     log(.verbose, msg: "Begin resolve cached object by component: \(component.info)", brace: .begin)
     defer { log(.verbose, msg: "End resolve cached object by component: \(component.info)", brace: .end) }
 
-    mutex.sync { _ = makeObject(by: component, use: nil) }
+    mutex.sync { _ = makeObject(by: component, use: nil, arguments: arguments) }
   }
   
-  func resolve<T>(type: T.Type = T.self, component: Component) -> T {
+  func resolve<T>(type: T.Type = T.self, component: Component, arguments: AnyArguments? = nil) -> T {
     let pType = ParsedType(type: type)
     log(.verbose, msg: "Begin resolve \(description(type: pType)) by component: \(component.info)", brace: .begin)
     defer { log(.verbose, msg: "End resolve \(description(type: pType)) by component: \(component.info)", brace: .end) }
     
-    return gmake(by: makeObject(by: component, use: nil))
+    return gmake(by: makeObject(by: component, use: nil, arguments: arguments))
   }
 
   /// Finds the most suitable components that satisfy the types.
@@ -170,7 +172,7 @@ class Resolver {
     mutex.sync { cache.containerStorage.clean() }
   }
   
-  private func make(by parsedType: ParsedType, with name: String?, from framework: DIFramework.Type?, use object: Any?) -> Any? {
+  private func make(by parsedType: ParsedType, with name: String?, from framework: DIFramework.Type?, use object: Any?, arguments: AnyArguments?) -> Any? {
     log(.verbose, msg: "Begin make \(description(type: parsedType))", brace: .begin)
     defer { log(.verbose, msg: "End make \(description(type: parsedType))", brace: .end) }
 
@@ -186,16 +188,16 @@ class Resolver {
         let saveGraph = cache.graph
 
         func makeDelayMaker(by parsedType: ParsedType, components: Components) -> Any? {
-          return delayMaker.init(container, { () -> Any? in
+          return delayMaker.init(container, { arguments -> Any? in
             return self.mutex.sync {
               // call `.value` into DI initialize.
               if saveGraph === self.cache.graph {
-                return self.make(by: parsedType, components: components, use: object)
+                return self.make(by: parsedType, components: components, use: object, arguments: arguments)
               }
               // Call `.value` firstly.
               if self.stack.isEmpty {
                 self.cache.graph = saveGraph.toStrongCopy()
-                return self.make(by: parsedType, components: components, use: object)
+                return self.make(by: parsedType, components: components, use: object, arguments: arguments)
               }
               // Call `.value` into DI initialize, and DI graph has lifetimes perContainer, perRun, single...
               // For this case need call provider on her Cache Graph.
@@ -206,7 +208,7 @@ class Resolver {
                 self.cache.graph.toWeak()
                 self.cache.graph = currentGraphCache
               }
-              return self.make(by: parsedType, components: components, use: object)
+              return self.make(by: parsedType, components: components, use: object, arguments: arguments)
             }
           })
         }
@@ -221,19 +223,19 @@ class Resolver {
         }
       }
 
-      return make(by: parsedType, components: components, use: object)
+      return make(by: parsedType, components: components, use: object, arguments: arguments)
     }
   }
 
   /// isMany for optimization
-  private func make(by parsedType: ParsedType, components: Components, use object: Any?) -> Any? {
+  private func make(by parsedType: ParsedType, components: Components, use object: Any?, arguments: AnyArguments?) -> Any? {
     if parsedType.hasMany {
       assert(nil == object, "Many injection not supported")
-      return components.sorted{ $0.order < $1.order }.compactMap{ makeObject(by: $0, use: nil) }
+      return components.sorted{ $0.order < $1.order }.compactMap{ makeObject(by: $0, use: nil, arguments: arguments) }
     }
 
     if let component = components.first, 1 == components.count {
-      return makeObject(by: component, use: object)
+      return makeObject(by: component, use: object, arguments: arguments)
     }
 
     if components.isEmpty {
@@ -247,7 +249,7 @@ class Resolver {
   }
   
   /// Super function
-  private func makeObject(by component: Component, use usingObject: Any?) -> Any? {
+  private func makeObject(by component: Component, use usingObject: Any?, arguments: AnyArguments?) -> Any? {
     log(.verbose, msg: "Found component: \(component.info)")
 
     let uniqueKey = component.info
@@ -282,12 +284,15 @@ class Resolver {
       return nil
     }
 
-    func getArgumentObject() -> Any? {
-      guard let extensions = container.extensionsContainer.optionalGet(by: component.info) else {
-        log(.error, msg: "Until get argument. Not found extensions for \(component.info)")
+    var componentArguments: Arguments?
+    func getArgumentObject(parameter: MethodSignature.Parameter) -> Any? {
+      componentArguments = componentArguments ?? arguments?.getArguments(for: component)
+
+      if componentArguments == nil {
+        log(.error, msg: "Get arguments for \(component.info) failed. Please specify arguments or remove dublicates for this type")
         return nil
       }
-      return extensions.getNextArg()
+      return componentArguments?.getArgument(for: parameter.parsedType.base.type)
     }
     
     func makeObject() -> Any? {
@@ -310,7 +315,8 @@ class Resolver {
           _ = use(signature: signature, usingObject: initializedObject)
         }
       }
-      
+
+      container.extensions.objectMaked?(uniqueKey, initializedObject)
       return initializedObject
     }
     
@@ -347,9 +353,9 @@ class Resolver {
         if parameter.parsedType.useObject {
           makedObject = usingObject
         } else if parameter.parsedType.arg {
-          makedObject = getArgumentObject()
+          makedObject = getArgumentObject(parameter: parameter)
         } else {
-          makedObject = make(by: parameter.parsedType, with: parameter.name, from: component.framework, use: nil)
+          makedObject = make(by: parameter.parsedType, with: parameter.name, from: component.framework, use: nil, arguments: arguments)
         }
         
         if nil != makedObject || parameter.parsedType.optional {
@@ -372,26 +378,32 @@ class Resolver {
       stack.removeLast()
     }
 
-    switch component.lifeTime {
-    case .single:
-      return makeObject(scope: Cache.single)
-    case .perRun(let referenceCounting):
-      switch referenceCounting {
-      case .weak: return makeObject(scope: Cache.weakPerRun)
-      case .strong: return makeObject(scope: Cache.strongPerRun)
+    func makeOrGetObject() -> Any? {
+      switch component.lifeTime {
+      case .single:
+        return makeObject(scope: Cache.single)
+      case .perRun(let referenceCounting):
+        switch referenceCounting {
+        case .weak: return makeObject(scope: Cache.weakPerRun)
+        case .strong: return makeObject(scope: Cache.strongPerRun)
+        }
+      case .perContainer(let referenceCounting):
+        switch referenceCounting {
+        case .weak: return makeObject(scope: cache.weakPerContainer)
+        case .strong: return makeObject(scope: cache.strongPerContainer)
+        }
+      case .objectGraph:
+        return makeObject(scope: cache.graph)
+      case .prototype:
+        return makeObject()
+      case .custom(let scope):
+        return makeObject(scope: scope)
       }
-    case .perContainer(let referenceCounting):
-      switch referenceCounting {
-      case .weak: return makeObject(scope: cache.weakPerContainer)
-      case .strong: return makeObject(scope: cache.strongPerContainer)
-      }
-    case .objectGraph:
-      return makeObject(scope: cache.graph)
-    case .prototype:
-      return makeObject()
-    case .custom(let scope):
-      return makeObject(scope: scope)
     }
+
+    let obj = makeOrGetObject()
+    container.extensions.objectResolved?(uniqueKey, obj)
+    return obj
   }
  
   private unowned let container: DIContainer