+++ Sources/Core/Internal/SpecificType.swift
@@ -2,7 +2,8 @@
 //  SpecificType.swift
 //  DITranquillity
 //
-//  Created by Ивлев Александр on 06/12/2018.
+//  Created by Alexander Ivlev on 06/12/2018.
+//  Copyright © 2018 Alexander Ivlev. All rights reserved.
 //
 
 protocol SpecificType {
@@ -24,29 +25,29 @@ protocol SpecificType {
 
     static func make(by obj: Any?) -> Self
 
-    init(_ container: DIContainer, _ factory: @escaping () -> Any?)
+    init(_ container: DIContainer, _ factory: @escaping (_ arguments: AnyArguments?) -> Any?)
 }
 
 extension SpecificType {
-    static var recursive: Bool { return true }
+  static var recursive: Bool { return true }
 
-    static var type: DIAType { return self }
-    static var tagType: DIAType { return self }
-    static var isSwiftType: Bool { return false }
+  static var type: DIAType { return self }
+  static var tagType: DIAType { return self }
+  static var isSwiftType: Bool { return false }
 
-    static var many: Bool { return false }
-    static var optional: Bool { return false }
-    static var delayed: Bool { return false }
-    static var inFramework: Bool { return false }
+  static var many: Bool { return false }
+  static var optional: Bool { return false }
+  static var delayed: Bool { return false }
+  static var inFramework: Bool { return false }
 
-    static var tag: Bool { return false }
-    static var arg: Bool { return false }
+  static var tag: Bool { return false }
+  static var arg: Bool { return false }
 
-    static var useObject: Bool { return false }
+  static var useObject: Bool { return false }
 
-    static func make(by obj: Any?) -> Self { return obj as! Self }
+  static func make(by obj: Any?) -> Self { return obj as! Self }
 
-    init(_ container: DIContainer, _ factory: @escaping () -> Any?) {
-        fatalError("Unsupport")
-    }
+  init(_ container: DIContainer, _ factory: @escaping (_ arguments: AnyArguments?) -> Any?) {
+    fatalError("Unsupport")
+  }
 }