+++ Sources/GraphAPI/DIGraph+Validate.swift
@@ -2,7 +2,7 @@
 //  DIGraph+Validate.swift
 //  DITranquillity
 //
-//  Created by Ивлев А.Е. on 02.07.2020.
+//  Created by Alexander Ivlev on 02.07.2020.
 //  Copyright © 2020 Alexander Ivlev. All rights reserved.
 //