+++ Sources/GraphAPI/DIVertex.swift
@@ -2,7 +2,7 @@
 //  DIVertex.swift
 //  DITranquillity
 //
-//  Created by Ивлев А.Е. on 06.07.2020.
+//  Created by Alexander Ivlev on 06.07.2020.
 //  Copyright © 2020 Alexander Ivlev. All rights reserved.
 //