+++ Tests/DITranquillityTest/DITranquillityTests_Extensions.swift
@@ -2,7 +2,7 @@
 //  DITranquillityTests_Extensions.swift
 //  DITranquillity
 //
-//  Created by Ивлев Александр Евгеньевич on 17/08/2018.
+//  Created by Alexander Ivlev on 17/08/2018.
 //  Copyright © 2018 Alexander Ivlev. All rights reserved.
 //
 
@@ -67,9 +67,7 @@ class DITranquillityTests_Extensions: XCTestCase {
       .injection{ $0.p4 = arg($1) }
       .injection(\.p5) { arg($0) }
 
-    container.extensions(for: Class.self)?.setArgs(111, 22.2, [1,2,3,4,5], ["H","e","l","l","o"])
-
-    let obj: Class = *container
+    let obj: Class = container.resolve(args: 111, 22.2, [1,2,3,4,5], ["H","e","l","l","o"])
 
     XCTAssertEqual(obj.p1, 111)
     XCTAssertEqual(obj.p2, "injection")
@@ -89,9 +87,8 @@ class DITranquillityTests_Extensions: XCTestCase {
 
     container.register(DepthClass.init)
 
-    container.extensions(for: Class.self)?.setArgs("injection", 22.2, [1,2,3,4,5], ["H","e","l","l","o"])
-
-    let obj: DepthClass = *container
+    let obj: DepthClass = container.resolve(
+      arguments: AnyArguments(for: Class.self, args: "injection", 22.2, [1,2,3,4,5], ["H","e","l","l","o"]))
 
     XCTAssertEqual(obj.c.p1, 111)
     XCTAssertEqual(obj.c.p2, "injection")
@@ -105,9 +102,7 @@ class DITranquillityTests_Extensions: XCTestCase {
 
     container.register{ OptionalClass(p1: arg($0), p2: $1, p3: arg($2)) }
 
-    container.extensions(for: OptionalClass.self)?.setArgs(nil, 22.2)
-
-    let obj: OptionalClass = *container
+    let obj: OptionalClass = container.resolve(args: nil, 22.2)
 
     XCTAssertNil(obj.p1)
     XCTAssertNil(obj.p2)
@@ -126,39 +121,19 @@ class DITranquillityTests_Extensions: XCTestCase {
     XCTAssertNil(obj.p3)
   }
 
-  func test05_ExtensionsNoRegister() {
+  func test05_ProtocolArguments() {
     let container = DIContainer()
 
-    let ext = container.extensions(for: Class.self)
-
-    XCTAssertNil(ext)
-  }
-
-  func test06_ExtensionsProtocolOneRegister() {
-    let container = DIContainer()
-
-   container.register{ Class(p1: $0, p2: arg($1), p3: arg($2)) }
-     .as(MyProtocol.self)
-
-    let ext = container.extensions(for: MyProtocol.self)
-
-    XCTAssertNotNil(ext)
-  }
-
-  func test06_ExtensionsMoreRegister() {
-     let container = DIContainer()
-
-    container.register{ OptionalClass(p1: arg($0), p2: $1, p3: arg($2)) }
+    container.register{ Class(p1: arg($0), p2: arg($1), p3: arg($2)) }
       .as(MyProtocol.self)
 
-    container.register{ Class(p1: $0, p2: arg($1), p3: arg($2)) }
-      .as(MyProtocol.self)
+    var arguments = AnyArguments()
+    arguments.addArgs(for: MyProtocol.self, args: 11, "test", 15.0)
+    let obj: Class = container.resolve(arguments: arguments)
 
-     let ext = container.extensions(for: MyProtocol.self)
-
-     XCTAssertNil(ext)
-   }
-
-
+    XCTAssertEqual(obj.p1, 11)
+    XCTAssertEqual(obj.p2, "test")
+    XCTAssertEqual(obj.p3, 15.0)
+  }
 }